Moved getMoodboards to its own action. Removed unused imports. Moodboard refresh when adding new moodboard is WIP.
